consolidate unparser integration tests
